requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
paddleocr==2.7.1.3
paddlepaddle==2.5.2
pillow
